+++ app/src/main/java/com/goalreminderbeta/sa/goalreminderbeta/goals/Goal.java
@@ -4,7 +4,7 @@ public interface Goal {
 
     boolean delete();
     Long getId();
-    int getDifferenceInDays();
+    double getDifferenceInDays();
     String getNameGoal();
     String getThemeCategory();
 }
