+++ app/src/main/java/com/goalreminderbeta/sa/goalreminderbeta/goals/LanguageCorrectionGoal.java
@@ -9,7 +9,7 @@ public class LanguageCorrectionGoal extends SugarRecord implements Goal{
 
     private Date currentDate;
     private Date goalDate;
-    private int differenceInDays;
+    private double differenceInDays;
     private String nameGoal;
     private String themeCategory;
 
@@ -22,7 +22,7 @@ public class LanguageCorrectionGoal extends SugarRecord implements Goal{
         this.nameGoal = nameGoal;
         this.themeCategory = themeCategory;if (currentDate!=null && goalDate!=null && (goalDate.getTime() - currentDate.getTime()) > 0){
             long milliseconds = goalDate.getTime() - currentDate.getTime();
-            this.differenceInDays = (int) TimeUnit.DAYS.convert(milliseconds, TimeUnit.MILLISECONDS);
+            this.differenceInDays = (double) TimeUnit.DAYS.convert(milliseconds, TimeUnit.MILLISECONDS);
         }
 
     }
@@ -43,11 +43,11 @@ public class LanguageCorrectionGoal extends SugarRecord implements Goal{
         this.goalDate = goalDate;
     }
 
-    public int getDifferenceInDays() {
+    public double getDifferenceInDays() {
         return differenceInDays;
     }
 
-    public void setDifferenceInDays(int differenceInDays) {
+    public void setDifferenceInDays(double differenceInDays) {
         this.differenceInDays = differenceInDays;
     }
