+++ app/src/main/java/com/goalreminderbeta/sa/goalreminderbeta/goals/ReadBookGoal.java
@@ -12,7 +12,7 @@ public class ReadBookGoal extends SugarRecord implements Goal {
     private Date currentDate;
     private Date goalDate;
     private String nameBook, nameAuthor;
-    private int differenceInDays;
+    private double differenceInDays;
     private String nameGoal;
     private String themeCategory;
 
@@ -29,7 +29,7 @@ public class ReadBookGoal extends SugarRecord implements Goal {
         this.themeCategory = themeCategory;
         if (currentDate!=null && goalDate!=null && (goalDate.getTime() - currentDate.getTime()) > 0){
             long milliseconds = goalDate.getTime() - currentDate.getTime();
-            this.differenceInDays = (int) TimeUnit.DAYS.convert(milliseconds, TimeUnit.MILLISECONDS);
+            this.differenceInDays = (double) TimeUnit.DAYS.convert(milliseconds, TimeUnit.MILLISECONDS);
         }
     }
 
@@ -42,11 +42,11 @@ public class ReadBookGoal extends SugarRecord implements Goal {
         this.themeCategory = themeCategory;
     }
 
-    public int getDifferenceInDays() {
+    public double getDifferenceInDays() {
         return differenceInDays;
     }
 
-    public void setDifferenceInDays(int differenceInDays) {
+    public void setDifferenceInDays(double differenceInDays) {
         this.differenceInDays = differenceInDays;
     }
